Move many tests out of WorkspaceApiServiceSpec to service-appropriate locations
